refactor: API domain별로 한 파일에 정리
